src/main/java/com/hernandes/andrade/fiap/hackatonfiasub/controller/AuthController.java
package com.hernandes.andrade.fiap.hackatonfiasub.controller;

import com.hernandes.andrade.fiap.hackatonfiasub.domain.User;
import com.hernandes.andrade.fiap.hackatonfiasub.usecase.AuthenticationService;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.mail.MessagingException;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/auth")
@RequiredArgsConstructor
@Tag(name = "Authentication", description = "Operations related to authentication")
public class AuthController {

    private final AuthenticationService authenticationService;

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.ACCEPTED)
    public ResponseEntity<?> register(
            @RequestBody @Valid RegistrationRequest registerRequest
    ) throws MessagingException {
        authenticationService.register(registerRequest);
        return ResponseEntity.accepted().build();
    }

    @PostMapping("/login")
    public ResponseEntity<LoginResponse> login(
            @RequestBody @Valid LoginRequest loginRequest
    ) {
        System.out.println("loginRequest = " + loginRequest);
        return ResponseEntity.ok(authenticationService.login(loginRequest));
    }

    @GetMapping("/activate-account")
    public ResponseEntity<?> activateAccount(
            @RequestParam String token
    ) throws MessagingException {
        authenticationService.activateAccount(token);
        return ResponseEntity.ok().build();
    }

    // Novo endpoint: Solicitação de recuperação de senha
    @PostMapping("/forgot-password")
    public ResponseEntity<?> forgotPassword(
            @RequestParam String email
    ) throws MessagingException {
        authenticationService.forgotPassword(email);
        return ResponseEntity.ok().build();
    }

    // Novo endpoint: Redefinição de senha
    @PostMapping("/reset-password")
    public ResponseEntity<?> resetPassword(
            @RequestParam String token,
            @RequestBody @Valid ResetPasswordRequest resetPasswordRequest
    ) {
        authenticationService.resetPassword(token, resetPasswordRequest);
        return ResponseEntity.ok().build();
    }

    @GetMapping("/hello")
    public String helloWord() {
        return "OLA MUNDO!";
    }

    /*private final UserRegistrationUseCase registrationUseCase;

    public AuthController(UserRegistrationUseCase registrationUseCase) {
        this.registrationUseCase = registrationUseCase;
    }

    @PostMapping("/register")
    public ResponseEntity<User> register(@RequestBody User user) {
        return ResponseEntity.ok(registrationUseCase.register(user));
    }*/
}
